% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crew.ssh-package.R
\docType{package}
\name{crew.ssh-package}
\alias{crew.ssh}
\alias{crew.ssh-package}
\title{crew.ssh: 'crew' Launcher Plugin for Running on a Remote System with 'ssh'}
\description{
In computationally demanding analysis projects, statisticians and data scientists may asynchronously deploy long-running tasks to remote systems by 'ssh'. This extends the 'crew' package with worker launcher plugins for remote systems by 'ssh'.
}
\author{
\strong{Maintainer}: Bill Denney \email{wdenney@humanpredictions.com} (\href{https://orcid.org/0000-0002-5759-428X}{ORCID})

Authors:
\itemize{
  \item Bill Denney \email{wdenney@humanpredictions.com} (\href{https://orcid.org/0000-0002-5759-428X}{ORCID})
}

}
\keyword{internal}
